Issue #96: Beginning work on a RPC client mode
At this point, the client can take a --client flag which will make
it pass all commands to an RPC server over the network rather than
processing them locally. Unfortunately, the client still starts all
of the normal network node services, so it connects to the network
and downloads the blockchain, etc. when all it should be doing is
connecting to the RPC server.