Refactor the push_block call tree
This tree needed to be cleaned up after my changes earlier. Some of the
functions were returning the fork_block_data of the second block from
the end of the longest fork, as a way to get the block_id of the end of
the fork. This was a dirty hack, and now the fork data of the last
block, as well as its block id, are returned together.