Don't munge memo on issue/create of assets, set issuer account
The scanners were replacing a meaningful memo on issue/create asset
transactions with a canned one. Fix this so that the memo is left alone
if it's already set.

Also, the from account was left empty on issue transactions, but the
from account must be the issuer (no one else can issue that asset) so we
can actually set it.