Resolve #571 and add regression test
Local accounts which have later had their names stolen on the blockchain
can now rename to a different name. In addition, a regression test has
been added to check that this renaming is allowed, while the registered
account still may not be renamed, nor may an account be renamed to a
name which is already taken on the chain.