Rework wallet use of active vs owner keys, update tests
The way I was adding the owner key to the wallet was inelegant and
hackish. I've changed the logic here to be much cleaner and more
flexible. In addition, the wallet was using the owner key for all
operations; now, it uses the active key for these operations.

I've also updated the tests, but I don't think the behavior is quite
correct yet.