Add config.json option to disable all things wallet
This is to support public-facing RPC servers which can provide a
blockchain interface to thin clients. Such an RPC server should not have
a wallet, nor should it be capable of having one, and all wallet-related
RPC calls should be disabled. This is now possible via the
wallet_enabled option in config.json.