Bring mail server into the fold
The mail server is now built into the client and can be enabled in the
config file. A new API section, mail, has been created, and a few calls
have been implemented therein.

This commit also introduces unit tests for the mail server (in
nathan_tests.cpp) and debugs the mail server so that the tests pass.