[hardfork] Add support for bid/ask in same trx
In the event a transaction specified a bid op and an ask op at the same
time it would fail to validate because of asset opperator += mismatch.

This will hard fork in the event that someone actually creates a
transaction that does this (ie: by using the market maker bot).

This expands the set of valid transactions, but does not invalidate past
transactions.