Add identity_verifier class
This class will manage the identity_verifier's role, which is really
quite simple. The verifier will store pending requests for verification,
serve them up for processing one at a time, and then store the result
until a client comes and requests it.

This commit is only the beginning of implementation; so far only some
basic database structures have been defined.