Further develop identity_verifier
Add methods to store and retrieve requests, write documentation on these
methods and the class as a whole.

Also, rework this index such that it contains all records, not just
recent ones, but remove the images from. The in-memory index should be
just the fields we might need to query by; it shouldn't contain the
large stuff. We can go to disk for that.