Issue #944 - Allow sender/receiver to cooperate
Under normal operations escrow is released from one party to another
and there is no need to use an escrow agent.  If an escrow agent is
used he can divide up the funds between sender/receiver; however,
sometimes an escrow agent is not available or the sender/receiver
can agree to other ways to split the funds.  If the balance is
signed by both the sender and receiver it has the same power
as if it was signed by just an escrow agent.