Tweak animations
Adjust spinner to use a stepped animation, add a pop transition
animation to the stackview.

As another point of interest, the spinner is now an animator, not an
animation, which means it does not run in the UI thread. This means that
the spinner will continue spinning even while the UI thread is blocked.
O.o