Refactor LightWallet
Change WelcomeLayout.qml to only deal with logins to existing accounts,
and make a dedicated OnboardingLayout.qml to handle the account creation
procedure. This simplifies WelcomeLayout.qml significantly as account
creation is significantly more complex than the nominal case of simply
logging in. Moreover, it eliminates the complexity of switching between
the two experiences in a single UI.